src/auth/auth_api.py
from fastapi import Security, HTTPException, status
from fastapi.security import APIKeyHeader
from src.classes.tipos import ConfigEnvSetings
import os
from dotenv import load_dotenv

load_dotenv()


API_KEY=ConfigEnvSetings.API_KEY
API_NAME=ConfigEnvSetings.API_NAME

api_key_header = APIKeyHeader(name=API_NAME, auto_error=False)

async def Auth_API_KEY(api_key: str = Security(api_key_header)):
    if not API_KEY:
        raise HTTPException(
            status_code=status.HTTP_500_INTERNAL_SERVER_ERROR,
            detail="API Key não configurada no servidor"
        )
    
    if not api_key:
        raise HTTPException(
            status_code=status.HTTP_401_UNAUTHORIZED,
            detail="API Key não fornecida",
            headers={"WWW-Authenticate": "ApiKey"},
        )
    
    if api_key != API_KEY:
        raise HTTPException(
            status_code=status.HTTP_403_FORBIDDEN,
            detail="API Key inválida",
            headers={"WWW-Authenticate": "ApiKey"},
        )
    
    return api_key

