requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
jinja2==3.1.2
aiofiles==23.2.1
requests==2.31.0
python-dotenv==1.0.0
itsdangerous==2.1.2
pydantic==2.5.0
pydantic-settings==2.1.0
openpyxl==3.1.2
email-validator==2.1.0

